+++ CHANGELOG.md
@@ -9,5 +9,7 @@
     + **ADDED** filter_fasta_by_tax_ids.py as main file. CLI implemented with docopt
     + **CHANGED** name of file to fafiti.py
     + **ADDED** folder for NCBI data (data/)
+    + **ADDED** function for downloading data by NCBI
+    
 
 ## Released